refact: add test3.h

+++ fec_codec.h
@@ -152,8 +152,8 @@ struct FecHeader {
 struct PacketLossStats {
     float    lossrate;                          ///< packet loss rate(in percentage) before recovery (-1 if unavailable)
     float    effective_lossrate;                ///< packet loss rate(in percentage) after recovery  (-1 if unavailable)
-    int32_t  missing_groups;                    ///< # missing groups (-1 if unavailable)
-    int32_t  loss_dist[kMaxContLossCount + 1];  ///< occurrences of successive packet losses
+    int64_t  missing_groups;                    ///< # missing groups (-1 if unavailable)
+    int64_t  loss_dist[kMaxContLossCount + 1];  ///< occurrences of successive packet losses
 };
 
 /**  fec packet memory layout
